Remove old specific-use pages
We should have all the forms and things on the index page.  Having
people jump around to tons of different pages for every single function
is way too complicated.

The idea was to be able to avoid javascript, but is it really that big
of a deal?  Not at all!  I'll just write some basic javascript to handle
alerting on success/failure.  Boom.  Easy.