Implement frontend functionality for email forms
These pages and scripts will properly display, submit, and report on
email form submissions.  They're still ugly as fuck, but they _work_.

Also, I tested the email functionality, and that works too!

What doesn't work at the moment is the actual form the email is supposed
to send you to, but that's just details...